refactor(v3/services): migrate ErrorTrackingService to services layer
Move ErrorTrackingService from exceptions/ to services/ to eliminate
circular dependency (exceptions → clients → exceptions).

Changes:
- Move error_tracking.py to services/error_tracking_service.py
- Move test files to tests/vibe3/services/
- Update all imports (top-level and lazy imports)
- Update all mock paths in tests
- Update services/__init__.py to export ErrorTrackingService
- Remove error_tracking comment from exceptions/__init__.py

Verification:
- All 72 tests pass
- mypy: no type errors
- ruff: all checks pass

Closes #1246
