+++ src/vibe3/orchestra/failed_gate.py
@@ -146,7 +146,7 @@ class FailedGate:
         Returns:
             GateResult with blocked=True if threshold reached
         """
-        from vibe3.exceptions.error_tracking import ErrorTrackingService
+        from vibe3.services.error_tracking_service import ErrorTrackingService
 
         log = logger.bind(domain="orchestra", action="error_threshold_check")
         log.debug("Checking error threshold")
@@ -257,7 +257,7 @@ class FailedGate:
             )
 
         # Also clear error log (use store-specific instance for consistency)
-        from vibe3.exceptions.error_tracking import ErrorTrackingService
+        from vibe3.services.error_tracking_service import ErrorTrackingService
 
         ErrorTrackingService.get_instance(store=self.store).clear(cleared_by, reason)
