+++ src/vibe3/services/error_tracking_service.py
