+++ src/vibe3/services/serve_status_service.py
@@ -9,9 +9,9 @@ from rich.console import Console
 from rich.table import Table
 
 from vibe3.config.orchestra_settings import load_orchestra_config
-from vibe3.exceptions.error_tracking import ErrorTrackingService
 from vibe3.models.orchestra_config import OrchestraConfig
 from vibe3.orchestra.failed_gate import FailedGate
+from vibe3.services.error_tracking_service import ErrorTrackingService
 from vibe3.utils.error_message_cleaner import (
     CODEAGENT_WRAPPER_RE,
     clean_error_message,
