+++ tests/vibe3/integration/test_error_severity_refactor.py
@@ -7,8 +7,8 @@ from pathlib import Path
 import pytest
 
 from vibe3.clients import SQLiteClient
-from vibe3.exceptions.error_tracking import ErrorTrackingService
 from vibe3.orchestra.failed_gate import FailedGate
+from vibe3.services.error_tracking_service import ErrorTrackingService
 
 
 @pytest.fixture(autouse=True)
