+++ tests/vibe3/services/test_dispatch_error_propagation.py
@@ -151,7 +151,9 @@ class TestFailedGateExecThreshold:
 
         # ErrorTrackingService is imported locally inside _check_error_threshold,
         # so we must patch the source module, not failed_gate.
-        with patch("vibe3.exceptions.error_tracking.ErrorTrackingService") as mock_ets:
+        with patch(
+            "vibe3.services.error_tracking_service.ErrorTrackingService"
+        ) as mock_ets:
             mock_instance = MagicMock()
             mock_instance.has_critical_error.return_value = False
             mock_instance.has_model_config_error.return_value = False
@@ -171,7 +173,9 @@ class TestFailedGateExecThreshold:
         """Single E_EXEC_* error below threshold should not trigger."""
         from vibe3.orchestra.failed_gate import FailedGate
 
-        with patch("vibe3.exceptions.error_tracking.ErrorTrackingService") as mock_ets:
+        with patch(
+            "vibe3.services.error_tracking_service.ErrorTrackingService"
+        ) as mock_ets:
             mock_instance = MagicMock()
             mock_instance.has_critical_error.return_value = False
             mock_instance.has_model_config_error.return_value = False
